Debugger ok, estava enviando sintese em vez de analise, subbanda 0 pnsr ruim

+++ Wavelets/WaveletHelper.h
@@ -10,7 +10,7 @@
 
 class WaveletHelper {
 public:
-    static vector<intMatrix> splitSubbands(int **InputImg, int ximg, int yimg, int nsubs);
+    static vector<intMatrix> splitSubbands(double **InputImg, int ximg, int yimg, int nsubs);
     static intMatrix quantize(const vector<intMatrix> &oldSubbands, const vector<vector<performance>> &performances, float lambda, unsigned bestCodebooks[NBANDS]);
 };
